feat(core): use double as default type

+++ src/IQFeed.CSharpApiClient.Examples/Examples/ConcurrentHistorical/ConcurrentHistoricalExample.cs
@@ -17,11 +17,11 @@ namespace IQFeed.CSharpApiClient.Examples.Examples.ConcurrentHistorical
         public string Name => typeof(ConcurrentHistoricalExample).Name;
         private const int NumberOfConcurrentClients = 15;
 
-        private readonly ConcurrentDictionary<string, List<DailyWeeklyMonthlyMessage<decimal>>> _dailyMessagesBySymbol;
+        private readonly ConcurrentDictionary<string, List<DailyWeeklyMonthlyMessage<double>>> _dailyMessagesBySymbol;
 
         public ConcurrentHistoricalExample() : base(LookupClientFactory.CreateNew(NumberOfConcurrentClients), NumberOfConcurrentClients)
         {
-            _dailyMessagesBySymbol = new ConcurrentDictionary<string, List<DailyWeeklyMonthlyMessage<decimal>>>();
+            _dailyMessagesBySymbol = new ConcurrentDictionary<string, List<DailyWeeklyMonthlyMessage<double>>>();
         }
 
         public void Run()
